refactor(components): refactor the App component and fix some bugs
